+++ ignite/cmd/scaffold_query.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/ignite/cli/ignite/pkg/cliui/clispinner"
+	"github.com/ignite/cli/ignite/pkg/cliui"
 	"github.com/ignite/cli/ignite/pkg/placeholder"
 )
 
@@ -38,8 +38,10 @@ func NewScaffoldQuery() *cobra.Command {
 func queryHandler(cmd *cobra.Command, args []string) error {
 	appPath := flagGetPath(cmd)
 
-	s := clispinner.New().SetText("Scaffolding...")
-	defer s.Stop()
+	session := cliui.New(cliui.StartSpinner())
+	defer session.End()
+
+	session.StartSpinner("Scaffolding...")
 
 	// Get the module to add the type into
 	module, err := cmd.Flags().GetString(flagModule)
@@ -83,15 +85,13 @@ func queryHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	s.Stop()
-
 	modificationsStr, err := sourceModificationToString(sm)
 	if err != nil {
 		return err
 	}
 
-	fmt.Println(modificationsStr)
-	fmt.Printf("\n🎉 Created a query `%[1]v`.\n\n", args[0])
+	session.Println(modificationsStr)
+	session.Printf("\n🎉 Created a query `%[1]v`.\n\n", args[0])
 
 	return nil
 }