+++ ignite/pkg/chaincmd/runner/runner.go
@@ -13,15 +13,13 @@ import (
 	"github.com/ignite/cli/ignite/pkg/chaincmd"
 	"github.com/ignite/cli/ignite/pkg/cmdrunner"
 	"github.com/ignite/cli/ignite/pkg/cmdrunner/step"
-	"github.com/ignite/cli/ignite/pkg/lineprefixer"
 	"github.com/ignite/cli/ignite/pkg/truncatedbuffer"
 )
 
 // Runner provides a high level access to a blockchain's commands.
 type Runner struct {
-	chainCmd                      chaincmd.ChainCmd
-	stdout, stderr                io.Writer
-	daemonLogPrefix, cliLogPrefix string
+	chainCmd       chaincmd.ChainCmd
+	stdout, stderr io.Writer
 }
 
 // Option configures Runner.
@@ -31,20 +29,6 @@ type Option func(r *Runner)
 func Stdout(w io.Writer) Option {
 	return func(runner *Runner) {
 		runner.stdout = w
-	}
-}
-
-// DaemonLogPrefix is a prefix added to app's daemon logs.
-func DaemonLogPrefix(prefix string) Option {
-	return func(runner *Runner) {
-		runner.daemonLogPrefix = prefix
-	}
-}
-
-// CLILogPrefix is a prefix added to app's cli logs.
-func CLILogPrefix(prefix string) Option {
-	return func(runner *Runner) {
-		runner.cliLogPrefix = prefix
 	}
 }
 
@@ -112,25 +96,17 @@ type runOptions struct {
 
 // run executes a command.
 func (r Runner) run(ctx context.Context, runOptions runOptions, stepOptions ...step.Option) error {
-	var (
-		// we use a truncated buffer to prevent memory leak
-		// this is because Stargate app currently send logs to StdErr
-		// therefore if the app successfully starts, the written logs can become extensive
-		errb = truncatedbuffer.NewTruncatedBuffer(runOptions.wrappedStdErrMaxLen)
+	// we use a truncated buffer to prevent memory leak
+	// this is because Stargate app currently send logs to StdErr
+	// therefore if the app successfully starts, the written logs can become extensive
+	errb := truncatedbuffer.NewTruncatedBuffer(runOptions.wrappedStdErrMaxLen)
 
-		// add optional prefixes to output streams.
-		stdout io.Writer = lineprefixer.NewWriter(r.stdout,
-			func() string { return r.daemonLogPrefix },
-		)
-		stderr io.Writer = lineprefixer.NewWriter(r.stderr,
-			func() string { return r.cliLogPrefix },
-		)
-	)
-
-	// Set standard outputs
+	stdout := r.stdout
 	if runOptions.stdout != nil {
 		stdout = io.MultiWriter(stdout, runOptions.stdout)
 	}
+
+	stderr := r.stderr
 	if runOptions.stderr != nil {
 		stderr = io.MultiWriter(stderr, runOptions.stderr)
 	}
@@ -142,7 +118,6 @@ func (r Runner) run(ctx context.Context, runOptions runOptions, stepOptions ...s
 		cmdrunner.DefaultStderr(stderr),
 	}
 
-	// Set standard input if defined
 	if runOptions.stdin != nil {
 		runnerOptions = append(runnerOptions, cmdrunner.DefaultStdin(runOptions.stdin))
 	}